+++ src/main/dominio/Obstaculo.java
@@ -3,7 +3,7 @@ package main.dominio;
 public class Obstaculo extends Entidad {
 	private boolean destructible;
 
-	public Obstaculo(double posX, double posY, boolean destructible) {
+	public Obstaculo(int posX, int posY, boolean destructible) {
 		super(posX, posY);
 		this.destructible = destructible;
 	}
